moved variables back
