+++ src/app/virtual-circuit/variables/variables.component.scss
